backend/requirements.txt
Django==3.2.7
djangorestframework
numpy
python-dateutil
pyyaml
scipy
uritemplate

